+++ src/app/_config/timerConfig.ts
