+++ src/app/timer/_components/TimerController.tsx
@@ -1,4 +1,4 @@
-import Button from "@/components/elements/Button";
+import Button from "@/app/_components/elements/Button";
 import { BsSkipEndFill } from "react-icons/bs";
 import { HiPause } from "react-icons/hi2";
 import { RiResetLeftFill } from "react-icons/ri";
